Add compatibility with older cmdline setup on console argument
Older systems featured full path to tty inside console=. This
is no longer valid and causes issues with fbcondecor. The new
expected format of console= is tty name with the directory part
stripped off: console=tty1 for example.
This patch has been in sabayon's genkernel for <2 years, but
should be tested out on non-standard environments.