add AUFS support inside linuxrc and initramfs.
If kernel is compiled with aufs support (which is not in the kernel
as of today) or aufs module is available, and cmdline contains "aufs"
the live system will boot using aufs.

This is just minimal support and requires to have aufs inside the
running kernel. Aufs module compilation might be added to the ebuild
in future.