Cleanup logic to find rootfs on ZFS
We no longer support root=ZFS. Only root=ZFS=$dataset is supported.
GRUB2 never used that syntax and users never liked it, so that should be
okay.

We also handle the case where the root pool is missing from the
cachefile or the cachefile itself is missing by informing users of how
to repair the system. We inform them rather than do it ourselves because
a step is required after the pool is imported and while we can do it for
them, the heuristics are unlikely to be right 100% of the time and we
really should not be making changes to the rootfs from the initramfs.

Signed-off-by: Richard Yao <ryao@gentoo.org>